refactor: use services
