+++ .gitignore
@@ -16,3 +16,4 @@ dist
 
 .DS_Store
 .vscode
+.zed
