--- .zed/settings.json
@@ -1,5 +1,0 @@
-// Folder-specific settings
-//
-// For a full list of overridable settings, and general information on folder-specific settings,
-// see the documentation: https://docs.zed.dev/configuration/configuring-zed#folder-specific-settings
-{}
